+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Added
 
+- [TD-6219] Generate embeddings for business concept versions
 - [TD-6901] Added validation to check if enriched text is empty for add to cache
 
 ## [7.5.0] 2025-040-30